* single component pattern for domain-workflows gates * clean up auth UI types and naming
