+++ src/components/app-nav-bar/app-nav-bar.styles.tsx
@@ -12,3 +12,11 @@ const cssStylesObj = {
 
 export const cssStyles: StyletronCSSObjectOf<typeof cssStylesObj> =
   cssStylesObj;
+
+export const errorSnackbarOverrides = {
+  Root: {
+    style: {
+      backgroundColor: '#c62828',
+    },
+  },
+};
