+++ src/components/app-nav-bar/use-auth-lifecycle.constants.ts
@@ -1,10 +1,2 @@
 export const LOGIN_ITEM = 'login';
 export const LOGOUT_ITEM = 'logout';
-
-export const ERROR_SNACKBAR_OVERRIDES = {
-  Root: {
-    style: {
-      backgroundColor: '#c62828',
-    },
-  },
-};
