+++ src/views/domain-workflows/domain-workflows.tsx
@@ -1,24 +1,14 @@
 'use client';
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { useSuspenseQuery } from '@tanstack/react-query';
-import dynamic from 'next/dynamic';
 
-import useSuspenseConfigValue from '@/hooks/use-config-value/use-suspense-config-value';
-import { type DescribeClusterResponse } from '@/route-handlers/describe-cluster/describe-cluster.types';
 import { type PublicAuthContext } from '@/utils/auth/auth-shared.types';
 import request from '@/utils/request';
 import { type DomainPageTabContentProps } from '@/views/domain-page/domain-page-content/domain-page-content.types';
 
-import isClusterAdvancedVisibilityEnabled from './helpers/is-cluster-advanced-visibility-enabled';
-
-const DomainWorkflowsBasic = dynamic(
-  () => import('@/views/domain-workflows-basic/domain-workflows-basic')
-);
-
-const DomainWorkflowsAdvanced = dynamic(
-  () => import('./domain-workflows-advanced/domain-workflows-advanced')
-);
+import DomainWorkflowsBasic from './domain-workflows-basic-lazy';
+import DomainWorkflowsClusterGate from './domain-workflows-cluster-gate';
 
 export default function DomainWorkflows(props: DomainPageTabContentProps) {
   const { data: authInfo } = useSuspenseQuery<PublicAuthContext>({
@@ -36,39 +26,4 @@ export default function DomainWorkflows(props: DomainPageTabContentProps) {
   }
 
   return <DomainWorkflowsClusterGate {...props} />;
-}
-
-function DomainWorkflowsClusterGate(props: DomainPageTabContentProps) {
-  const { data: clusterInfo } = useSuspenseQuery<DescribeClusterResponse>({
-    queryKey: ['describeCluster', props.cluster],
-    queryFn: () =>
-      request(`/api/clusters/${props.cluster}`).then((res) => res.json()),
-    retry: false,
-  });
-
-  const isAdvancedVisibilityEnabled = useMemo(() => {
-    return isClusterAdvancedVisibilityEnabled(clusterInfo);
-  }, [clusterInfo]);
-
-  if (!isAdvancedVisibilityEnabled) {
-    return (
-      <DomainWorkflowsBasic domain={props.domain} cluster={props.cluster} />
-    );
-  }
-
-  return <DomainWorkflowsAdvancedGate {...props} />;
-}
-
-function DomainWorkflowsAdvancedGate(props: DomainPageTabContentProps) {
-  const { data: isNewWorkflowsListEnabled } = useSuspenseConfigValue(
-    'WORKFLOWS_LIST_ENABLED'
-  );
-
-  return (
-    <DomainWorkflowsAdvanced
-      domain={props.domain}
-      cluster={props.cluster}
-      isNewWorkflowsListEnabled={isNewWorkflowsListEnabled}
-    />
-  );
 }
